requirements.txt
cryptography~=41.0.1
pynacl==1.5.0